LayoutTests/fast/dom/CSSStyleDeclaration/transition-property-names.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
'This test checks that CSS property names work round trip in the transition property.'
);

var element = document.createElement('a');

element.style.webkitTransitionProperty = "height";
shouldBe("element.style.webkitTransitionProperty", "'height'");

element.style.webkitTransitionProperty = "opacity";
shouldBe("element.style.webkitTransitionProperty", "'opacity'");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
