third_party/blink/web_tests/fast/dom/CSSStyleDeclaration/transition-property-names-expected.txt

This test checks that CSS property names work round trip in the transition property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS element.style.webkitTransitionProperty is 'height'
PASS element.style.webkitTransitionProperty is 'opacity'
PASS successfullyParsed is true

TEST COMPLETE
